backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.8.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-dotenv>=1.0.0
